Fix Bugs

+++ connectDB.sh
@@ -3,60 +3,39 @@
 export pat="^[a-zA-Z]+[a-zA-Z0-9[:space:]]*"
 
 function showTables {
-            zenity --info \
-                        --title "Info Message" \
-                        --width 500 \
-                        --height 100 \
-                        --text "Tables Found: 
-`ls | awk 'BEGIN{FS=" "} { print "- "$1 }'`"
-            #ls | awk 'BEGIN{FS=" "} { print "- "$1 }';
-         }
+    res=$(ls | awk 'BEGIN{FS=" "} { print "- "$1 }')
+    if [[ $res == "" ]]; then
+        echo "No Tables found"
+    else
+        echo $res
+    fi
+    showMenu
+}
 
 function dropTable {
-    name=$(zenity --entry \
-        --width 500 \
-        --title "check user" \
-        --text "`-e` Enter Table Name: " );
-  #echo -e "Enter Table Name: \c"
-  #read name
-  rm $name 
-  rm .$name 
-  if [[ $? == 0 ]]
-  then
-    zenity --info \
-        --title "Info Message" \
-        --width 500 \
-        --height 100 \
-        --text "Table Dropped Successfully"
-    #echo "Table Dropped Successfully"
-  else
-    zenity --info \
-        --title "Info Message" \
-        --width 500 \
-        --height 100 \
-        --text "Error Dropping Table $tName"
-    #echo "Error Dropping Table $tName"
-  fi
+    read -p "Enter Table Name: " name
+    if [ -f "$name" ]; then
+        read -p "You are going to drop table $name? Are you sure? [Y/N]: " check
+        if [[ $check = "Y" || $check = "y" ]]; then
+            rm $name
+            rm .$name
+            echo "Table Dropped Successfully"
+        fi
+    else
+        echo "$name table is not existed!"
+    fi
+    showMenu
+
 }
 
 function showMenu() {
-    choice=$(zenity --list \
-      --column "Select Menu" \
-      createTable \
-      listTable \
-      dropTable \
-      insertTable \
-      selectTable \
-      removeFromTable \
-      updateTable \
-      "Go back to main menu"
-      )
-    #select choice in createTable listTable dropTable insertTable selectTable removeFromTable updateTable "Go back to main menu"; do
+    select choice in createTable listTable dropTable insertTable selectTable removeFromTable updateTable "Go back to main menu"; do
         case $choice in
         createTable)
-            . createTable.sh
+            . ../../createTable.sh
             ;;
         listTable)
+            pwd
             echo "listTable"
             showTables
             showMenu
@@ -64,21 +43,21 @@ function showMenu() {
         dropTable)
             echo "dropTable"
             dropTable
-            showMenu
             ;;
         insertTable)
-            . insertTable.sh
+            . ../../insertTable.sh
             ;;
         selectTable)
             echo "selectTable"
-            . selectTable.sh
+            . ../../selectTable.sh
             ;;
         removeFromTable)
             echo "removeFromTable"
-            . deleteDataFromTable.sh
+            . ../../deleteDataFromTable.sh
             ;;
         updateTable)
             echo "updateTable"
+            . ../../updateTable.sh
             ;;
         "Go back to main menu")
             cd ../..
@@ -86,16 +65,11 @@ function showMenu() {
             break
             ;;
         *)
-            zenity --info \
-                --title "Info Message" \
-                --width 500 \
-                --height 100 \
-                --text "wrong input select again"
-            #echo "wrong input select again"
+            echo "wrong input select again"
             showMenu
             ;;
         esac
-    #done
+    done
 }
 
 showMenu
